pyproject.toml
[project]
name = "realtimeaudio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "streamlit>=1.39.0",
    "sounddevice>=0.5.0",
    "numpy>=2.1.1",
    "google-cloud-speech>=2.27.0",
    "google-cloud-aiplatform>=1.69.0",
    "vertexai>=1.69.0",
    "streamlit-webrtc>=0.47.9",
    "av>=12.3.0",
    "websockets>=13.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
